refactored "undo" in data entry window
Undo is now a function with an undo stack but only 1 action
will be able to be undone per button call. This is enough for
this version and can be expanded later if needed.
Also added some error handling for undoing when there is
nothing to undo or the undo is impossible because its more than
1 action of the same button ago.